+++ backend/modules/blog/actions/edit_category.php
@@ -70,10 +70,11 @@ class BackendBlogEditCategory extends BackendBaseActionEdit
 	protected function parse()
 	{
 		parent::parse();
+
 		$this->tpl->assign('item', $this->record);
 
 		// delete allowed?
-		$this->tpl->assign('deleteAllowed', BackendBlogModel::deleteCategoryAllowed($this->id));
+		$this->tpl->assign('showDeleteCategory', BackendBlogModel::deleteCategoryAllowed($this->id) && BackendModel::createURLForAction('delete_category'));
 	}
 
 	/**