+++ backend/modules/extensions/actions/detail_module.php
@@ -157,13 +157,15 @@ class BackendExtensionsDetailModule extends BackendBaseActionIndex
 	/**
 	 * Parse.
 	 */
-	private function parse()
+	protected function parse()
 	{
+		parent::parse();
+
 		// assign module data
 		$this->tpl->assign('name', $this->currentModule);
 		$this->tpl->assign('warnings', $this->warnings);
 		$this->tpl->assign('information', $this->information);
-		$this->tpl->assign('isInstallable', !BackendExtensionsModel::isModuleInstalled($this->currentModule));
+		$this->tpl->assign('showInstallModule', !BackendExtensionsModel::isModuleInstalled($this->currentModule) && BackendAuthentication::isAllowedAction('install_module'));
 
 		// data grids
 		$this->tpl->assign('dataGridEvents', (isset($this->dataGridEvents) && $this->dataGridEvents->getNumResults() > 0) ? $this->dataGridEvents->getContent() : false);