+++ backend/modules/extensions/actions/install_module.php
@@ -1,5 +1,12 @@
 <?php
 
+/*
+ * This file is part of Fork CMS.
+ *
+ * For the full copyright and license information, please view the license
+ * file that was distributed with this source code.
+ */
+
 /**
  * This is the module install-action.
  * It will install the module given via the "module" GET parameter.