+++ backend/modules/mailmotor/actions/edit_mailing_campaign.php
@@ -76,7 +76,6 @@ class BackendMailmotorEditMailingCampaign extends BackendBaseActionEdit
 	 */
 	protected function parse()
 	{
-		// call parent
 		parent::parse();
 
 		// assign the active record and additional variables