+++ backend/modules/pages/actions/index.php
@@ -64,18 +64,22 @@ class BackendPagesIndex extends BackendBaseActionIndex
 		// disable paging
 		$this->dgDrafts->setPaging(false);
 
-		// set colum URLs
-		$this->dgDrafts->setColumnURL('title', BackendModel::createURLForAction('edit') . '&amp;id=[id]&amp;draft=[revision_id]');
-
 		// set column functions
 		$this->dgDrafts->setColumnFunction(array('BackendDataGridFunctions', 'getUser'), array('[user_id]'), 'user_id', true);
 		$this->dgDrafts->setColumnFunction(array('BackendDataGridFunctions', 'getLongDate'), array('[edited_on]'), 'edited_on');
 
-		// add edit column
-		$this->dgDrafts->addColumn('edit', null, BL::lbl('Edit'), BackendModel::createURLForAction('edit') . '&amp;id=[id]&amp;draft=[revision_id]', BL::lbl('Edit'));
-
 		// set headers
 		$this->dgDrafts->setHeaderLabels(array('user_id' => SpoonFilter::ucfirst(BL::lbl('By')), 'edited_on' => SpoonFilter::ucfirst(BL::lbl('LastEdited'))));
+
+		// check if allowed to edit
+		if(BackendAuthentication::isAllowedAction('edit', $this->getModule()))
+		{
+			// set column URLs
+			$this->dgDrafts->setColumnURL('title', BackendModel::createURLForAction('edit') . '&amp;id=[id]&amp;draft=[revision_id]');
+
+			// add edit column
+			$this->dgDrafts->addColumn('edit', null, BL::lbl('Edit'), BackendModel::createURLForAction('edit') . '&amp;id=[id]&amp;draft=[revision_id]', BL::lbl('Edit'));
+		}
 	}
 
 	/**
@@ -96,14 +100,18 @@ class BackendPagesIndex extends BackendBaseActionIndex
 		$this->dgRecentlyEdited->setColumnFunction(array('BackendDataGridFunctions', 'getUser'), array('[user_id]'), 'user_id');
 		$this->dgRecentlyEdited->setColumnFunction(array('BackendDataGridFunctions', 'getTimeAgo'), array('[edited_on]'), 'edited_on');
 
-		// set column URL
-		$this->dgRecentlyEdited->setColumnUrl('title', BackendModel::createURLForAction('edit') . '&amp;id=[id]', BL::lbl('Edit'));
-
-		// add column
-		$this->dgRecentlyEdited->addColumn('edit', null, BL::lbl('Edit'), BackendModel::createURLForAction('edit') . '&amp;id=[id]', BL::lbl('Edit'));
-
 		// set headers
 		$this->dgRecentlyEdited->setHeaderLabels(array('user_id' => SpoonFilter::ucfirst(BL::lbl('By')), 'edited_on' => SpoonFilter::ucfirst(BL::lbl('LastEdited'))));
+
+		// check if allowed to edit
+		if(BackendAuthentication::isAllowedAction('edit', $this->getModule()))
+		{
+			// set column URL
+			$this->dgRecentlyEdited->setColumnUrl('title', BackendModel::createURLForAction('edit') . '&amp;id=[id]', BL::lbl('Edit'));
+
+			// add column
+			$this->dgRecentlyEdited->addColumn('edit', null, BL::lbl('Edit'), BackendModel::createURLForAction('edit') . '&amp;id=[id]', BL::lbl('Edit'));
+		}
 	}
 
 	/**
@@ -121,14 +129,19 @@ class BackendPagesIndex extends BackendBaseActionIndex
 	/**
 	 * Parse the datagrid and the reports
 	 */
-	private function parse()
+	protected function parse()
 	{
+		parent::parse();
+
 		// parse dgRecentlyEdited
 		$this->tpl->assign('dgRecentlyEdited', ($this->dgRecentlyEdited->getNumResults() != 0) ? $this->dgRecentlyEdited->getContent() : false);
 		$this->tpl->assign('dgDrafts', ($this->dgDrafts->getNumResults() != 0) ? $this->dgDrafts->getContent() : false);
 
 		// parse the tree
 		$this->tpl->assign('tree', BackendPagesModel::getTreeHTML());
+
+		// check if allowed to add
+		$this->tpl->assign('showAdd', BackendAuthentication::isAllowedAction('add', $this->getModule()));
 
 		// open the tree on a specific page
 		if($this->getParameter('id', 'int') !== null) $this->tpl->assign('openedPageId', $this->getParameter('id', 'int'));