+++ frontend/modules/mailmotor/engine/helper.php
@@ -1,5 +1,12 @@
 <?php
 
+/*
+ * This file is part of Fork CMS.
+ *
+ * For the full copyright and license information, please view the license
+ * file that was distributed with this source code.
+ */
+
 /**
  * In this file we store all generic functions that we will be using to communicate with CampaignMonitor
  *
@@ -123,7 +130,7 @@ class FrontendMailmotorCMHelper
 
 		// set groupID
 		$groupId = !empty($groupId) ? $groupId : FrontendMailmotorModel::getDefaultGroupID();
-		
+
 		// get campaign monitor list id
 		$listId = self::getCampaignMonitorID('list', $groupId);